# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Analysis Canvas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/users/{user_id}/status": {
      "patch": {
        "tags": [
          "access-control"
        ],
        "summary": "Update Account Status",
        "operationId": "update_account_status_api_admin_users__user_id__status_patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Account Status Api Admin Users  User Id  Status Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/global-admin": {
      "patch": {
        "tags": [
          "access-control"
        ],
        "summary": "Update Global Admin",
        "operationId": "update_global_admin_api_admin_users__user_id__global_admin_patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalAdminUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Global Admin Api Admin Users  User Id  Global Admin Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/requests": {
      "get": {
        "summary": "Get Requests",
        "operationId": "get_requests_api_projects__project_id__requests_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Requests Api Projects  Project Id  Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Request",
        "operationId": "create_request_api_projects__project_id__requests_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Request Api Projects  Project Id  Requests Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requests/{request_id}/assignee": {
      "patch": {
        "summary": "Reassign Request",
        "operationId": "reassign_request_api_requests__request_id__assignee_patch",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssigneeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reassign Request Api Requests  Request Id  Assignee Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analysis-runs/{run_id}/review-items": {
      "get": {
        "summary": "List Review Items",
        "operationId": "list_review_items_api_analysis_runs__run_id__review_items_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Review Items Api Analysis Runs  Run Id  Review Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Review Item",
        "operationId": "create_review_item_api_analysis_runs__run_id__review_items_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Review Item Api Analysis Runs  Run Id  Review Items Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review-items/{annotation_id}": {
      "patch": {
        "summary": "Update Review Item",
        "operationId": "update_review_item_api_review_items__annotation_id__patch",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Review Item Api Review Items  Annotation Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountStatusUpdate": {
        "properties": {
          "account_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "SUSPENDED"
            ],
            "title": "Account Status"
          },
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "account_status",
          "expected_updated_at",
          "reason"
        ],
        "title": "AccountStatusUpdate"
      },
...
      "AnalysisRequestCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Title"
          },
          "owner_user_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Owner User Id"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "due_in_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Due In Days",
            "default": 7
          },
          "overall_note": {
            "type": "string",
            "maxLength": 500,
            "title": "Overall Note",
            "default": ""
          },
          "source_type": {
            "type": "string",
            "enum": [
              "EXTERNAL_SYSTEM",
              "DEPARTMENT_HEAD"
            ],
            "title": "Source Type"
          },
          "source_reference": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Source Reference"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          },
          "request_type_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_-]*$",
            "title": "Request Type Id",
            "default": "design-reliability-validation"
          },
          "request_type_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Request Type Version",
            "default": 1
          },
          "assigned_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned By"
          }
        },
        "type": "object",
        "required": [
          "title",
          "owner_user_id",
          "source_type",
          "source_reference"
        ],
        "title": "AnalysisRequestCreate"
      },
...
      "AssigneeUpdate": {
        "properties": {
          "owner_user_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Owner User Id"
          }
        },
        "type": "object",
        "required": [
          "owner_user_id"
        ],
        "title": "AssigneeUpdate"
      },
...
      "GlobalAdminUpdate": {
        "properties": {
          "is_global_admin": {
            "type": "boolean",
            "title": "Is Global Admin"
          },
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "is_global_admin",
          "expected_updated_at",
          "reason"
        ],
        "title": "GlobalAdminUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewItemCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Body"
          },
          "variable_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Variable Key"
          },
          "time_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Value"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "NODE",
                  "ELEMENT"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Type"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "review_status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_REVIEW",
              "RESOLVED"
            ],
            "title": "Review Status",
            "default": "OPEN"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "title": "ReviewItemCreate"
      },
      "ReviewItemUpdate": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "review_status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_REVIEW",
              "RESOLVED"
            ],
            "title": "Review Status"
          }
        },
        "type": "object",
        "required": [
          "review_status"
        ],
        "title": "ReviewItemUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}